roles

+++ app/Http/Controllers/Api/RoleController.php
@@ -3,40 +3,75 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Http\Resources\RoleResource;
 use App\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Symfony\Component\HttpFoundation\Response;
 
 class RoleController extends Controller
 {
     public function index()
     {
-        return Role::all();
-    }
+        Gate::authorize('view', 'roles');
 
-    public function store(Request $request)
-    {
-        $role = Role::create($request->only('name'));
-
-        return response($role, Response::HTTP_CREATED);
+        return RoleResource::collection(Role::all());
     }
 
     public function show($id)
     {
-        return Role::find($id);
+        Gate::authorize('view', 'roles');
+
+        return new RoleResource(Role::find($id));
+    }
+
+    public function store(Request $request)
+    {
+        Gate::authorize('edit', 'roles');
+
+        $role = Role::create($request->only('name'));
+
+        if ($permissions = $request['permissions']) {
+            foreach ($permissions as $permission_id) {
+                DB::table('permission_role')->insert([
+                    'role_id' => $role->id,
+                    'permission_id' => $permission_id,
+                ]);
+            }
+        }
+
+        return response(new RoleResource($role), Response::HTTP_CREATED);
     }
 
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::find($id);
 
         $role->update($request->only('name'));
 
-        return response($role, Response::HTTP_ACCEPTED);
+        DB::table('permission_role')->where('role_id', $role->id)->delete();
+
+        if ($permissions = $request['permissions']) {
+            foreach ($permissions as $permission_id) {
+                DB::table('permission_role')->insert([
+                    'role_id' => $role->id,
+                    'permission_id' => $permission_id,
+                ]);
+            }
+        }
+
+        return response(new RoleResource($role), Response::HTTP_ACCEPTED);
     }
 
     public function destroy($id)
     {
+        Gate::authorize('edit', 'roles');
+
+        DB::table('permission_role')->where('role_id', $id)->delete();
+
         Role::destroy($id);
 
         return response(null, Response::HTTP_NO_CONTENT);
